Clarify integer widths in CBookmark::Distance()
m_iFrame - other.m_iFrame is intended to produce a negative result.
In the old code, it produced a very large unsigned integer (bad),
but it got casted to a negative signed integer upon * or %,
so the code was merely unclear but not wrong.
Anyway I've cleaned it up (at the cost of verbosity).